added stateful support in kube api
