+++ requirements.txt
@@ -1,3 +1,4 @@
 kubernetes==26.1.0
 Flask==2.3.2
 gunicorn==20.1.0
+awscli
